+++ src/sub-clients/adminLinks/model/AdminLinks.ts
@@ -39,4 +39,5 @@ export interface AdminLinks {
   ingestVideoStatuses: LinkEntity;
   contractLegalRestrictions: LinkEntity;
   getCaptions: LinkEntity;
+  contentWarnings: LinkEntity;
 }