+++ src/sub-clients/contentWarnings/client/ContentWarningsClient.ts
@@ -1,0 +1,5 @@
+import { ContentWarning } from '../model/ContentWarning';
+
+export interface ContentWarningsClient {
+  getAll(): Promise<ContentWarning[]>;
+}